+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Explicitly allow reading the `fhirUser` resource when scope and launch context ID are present (RFS-309)
+
 ## [0.2.0] - 2026-07-13
 
 ### Fixed